requirements.txt
fastapi==0.115.12
faster-whisper==1.1.1
fugashi[unidic]==1.4.0
jamdict==0.1a11.post2
openai==1.76.2
pydantic==2.11.4
python-dotenv==1.1.0
srt==3.5.3
python-multipart==0.0.20
aiofiles==24.1.0
uvicorn==0.34.2
pydantic[email]
modal==0.75.6
genanki==0.13.1